Item Creation
Modified apps.py removed use of reserved word id by putting in a _.

function _is_valid_data now returns false if passed (i know this is
counter intuitive atm, but if it doesn't pass it returns the error
message, which will eval to true...

Some changes might be needed to the item is_valid function.

Minor bug fixes in the NEPBoundField funcitons that fix a syntax error
in angular when a number is used as the field name.

Item form will now build it's fields from the item type.