Item Update/Create and datatables
Worked on the item editing and the datatable for items.

Lots of work done on angular to get paramatarized views working without
generating states with django templates. This took most of my time ;(.

Edit view should be functional once Herco fixes his create item
function.

Controllers created for list. New list controllers inherit from
'ListControllor' and then define a url, a create state, a edit state and
a delete state.

url is what the controllor calls to get paginated lists from.
create/edit and delete states are the views to be renedered when a
object is to be cruded.