Menu framework improvments
Menus now run of the apps.get_app_config, instead of INSTALLED_APPS
setting.

I also added menu and state models so that menus can be configured via
the django admin.

The menus will now be built from both the menu.py files and the mysql
database.

Menu children are now stored in a dictionary with the menu.text as the
key. So menus can't share names with siblings.