Realtime Dynamic Menus (WOOOO)
Menus will now update in realtime via websockets!

NOTES:
Websockets are an actual socket. I think we should keep this to one
connection and use the built-in stream functionality that
django-channels provides. The sockets all connect to the same url. The
goal of the messages are then identified using the streams.

BACKEND:
Using django channels, a message is sent over a websocket whenever an
instance of a model is changed.  This is sent over the "menu" stream.
This will only work if it goes through the proper saving mechanism. If
you use the queryset directly django does not fire all events.

FRONTEND:
Using the javascript library that django-channels provides I created a
websocket that will delegate events. It uses the stream id. In this case
"menu".