NepList Initial
Added checklists that can be linked to checks. File uploading works with
this version of neptune. Maybe considering porting that to the main
neptune branch.

Checks can be viewed from the front-end.

# TODO
Link the cron up so that teh worklist will create a new one once the old
one is succesfully saved. Also limit display of list so that only lists
for the correct date are displayed.